Added function to filter caliop 1km data using optical depth info from caliop 5km data. Remove the top of the clouds down to the OPTICAL_DETECTION_LIMIT. Function name: use5km_remove_thin_clouds_from_1km.
Tested to use middel of cloud instead of the top for clouds deeper than 1km.
Code not used, included if we want it later.

Added posibility to sort out calipso pixels where the variation is high in the 1km pixels mathing a 5km pixel.
Code note used now. Included if we want to develop and use it later.