Bugfix matching can now handle nodata in lat/lon.
The indices delivered by get_neighbour_info are refering to valid pixels.
For an avhrr with 2 lines of nodata,
this means index 0 is actually referring to pixel ~819.
And linenumber 0 should have been line number 3 and so on.
Causing the matched pixels to not match at all.

This matching was always used for cph and lwp.
However in locally recieved data nodata in lat/lon is rare.